Add files via upload
It is a project based on how to manage the login of users and admin ,and also to manage the number of beds available, In this project the user and admin can able to login if the user didn't have account then he able to register a new account. Where the admin can only able to Login and manage the user's login.